backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
psycopg2-binary
google-auth-oauthlib
google-api-python-client
bcrypt==3.2.0
passlib[bcrypt]
python-jose[cryptography]